Break Login and Register out into Container/Presentational components
