+++ client/src/components/App/index.js
@@ -2,7 +2,8 @@ import React from "react"
 import { Route, Switch, Redirect } from "react-router-dom"
 import { connect } from "react-redux"
 
-import { Register, Login } from "../Auth"
+import Login from "../Containers/Login"
+import Register from "../Containers/Register"
 import AppContainer from "../AppContainer"
 import FakeLogin from "../FakeLogin"
 import LandingPage from "../Pages/LandingPage"
