--- client/src/components/Auth/index.js
@@ -1,2 +1,0 @@
-export { default as Register } from "./Register"
-export { default as Login } from "./Login"
