+++ client/src/components/Containers/Banner/index.js
