+++ client/src/components/Presentational/Login/styles.js
@@ -1,0 +1,13 @@
+import styled from "styled-components"
+
+export const LoginStyles = styled.div`
+  height: 100%;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  form {
+    width: 300px;
+  }
+`
