+++ client/src/components/Presentational/Register/styles.js
@@ -1,0 +1,14 @@
+import styled from "styled-components"
+
+export const RegisterStyles = styled.div`
+  /* margin-top: ${props => "-" + props.theme.navHeight}; */
+  height: 100%;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  form {
+    width: 300px;
+  }
+`
